requirements.txt
streamlit==1.28.2
pandas==2.1.4
requests==2.31.0
openpyxl==3.1.2
xlrd==2.0.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
python-dotenv==1.0.0
